+++ src/main/java/hu/ibello/model/BrowserMemoryUsage.java
@@ -1,20 +1,57 @@
+/*
+ * Ark-Sys Kft. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package hu.ibello.model;
 
 import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlType;
 
 @XmlRootElement(name="browser-memory")
-public class BrowserMemoryUsage extends JavaMemoryUsage {
+@XmlType(propOrder={"windowId", "total", "used"})
+public class BrowserMemoryUsage {
 
-	private String name;
+	private String WindowId;
+	private Long total;
+	private Long used;
 	
-	public String getName() {
-		return name;
+	public String getWindowId() {
+		return WindowId;
+	}
+	
+	@XmlAttribute(name="window-id")
+	public void setWindowId(String windowId) {
+		WindowId = windowId;
+	}
+	
+	public Long getTotal() {
+		return total;
 	}
 	
 	@XmlAttribute
-	public void setName(String name) {
-		this.name = name;
+	public void setTotal(Long total) {
+		this.total = total;
 	}
 	
+	public Long getUsed() {
+		return used;
+	}
+	
+	@XmlAttribute
+	public void setUsed(Long used) {
+		this.used = used;
+	}
+
 }