+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New `testData().fromBinary(String)` method for loading binary test data
 - New `testData().fromProperties(String).loadString()` and `testData().fromJson(Class).loadString()` methods for loading complex test data as text
 - New `testData().fromProperties(String).openStream()` and `testData().fromJson(Class).openStream()` methods for loading complex test data as stream
+- New `testData().fromProperties(String).withCharset(Charset)` and `testData().fromJson(Class).withCharset(Charset)` methods to specify the character set used during test data loading
 
 ## 1.11.0